+++ src/mocks.specUtil.ts
@@ -39,6 +39,7 @@ export function createTestConfiguration(overrides: Partial<Configuration> = {}):
     telemetrySampleRate: 100,
     defaultPrivacyLevel: 'mask',
     allowedWebViewHosts: [],
+    correctPrewarmedViewTimings: true,
     ...overrides,
   };
 }